Added LastTrade

+++ CrypConnect/Exchanges/Kucoin/KucoinJson.cs
@@ -9,19 +9,19 @@ namespace CrypConnect.Exchanges.Kucoin
       public string coinType { get; set; }
       public bool trading { get; set; }
       public string symbol { get; set; }
-      public float lastDealPrice { get; set; }
-      public float buy { get; set; }
-      public float sell { get; set; }
-      public float change { get; set; }
+      public double lastDealPrice { get; set; }
+      public double buy { get; set; }
+      public double sell { get; set; }
+      public double change { get; set; }
       public string coinTypePair { get; set; }
       public int sort { get; set; }
-      public float feeRate { get; set; }
-      public float volValue { get; set; }
-      public float high { get; set; }
+      public double feeRate { get; set; }
+      public double volValue { get; set; }
+      public double high { get; set; }
       public long datetime { get; set; }
-      public float vol { get; set; }
-      public float low { get; set; }
-      public float changeRate { get; set; }
+      public double vol { get; set; }
+      public double low { get; set; }
+      public double changeRate { get; set; }
     }
 
     public bool success { get; set; }
